# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UnivMarket — Algerian University Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services with filters
      parameters:
        - name: wilaya_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: seller_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, banned]
        - name: featured
          in: query
          schema:
            type: boolean
        - name: min_price
          in: query
          schema:
            type: ["number", "null"]
        - name: max_price
          in: query
          schema:
            type: ["number", "null"]
        - name: min_rating
          in: query
          schema:
            type: ["number", "null"]
        - name: price_type
          in: query
          schema:
            type: string
            enum: [fixed, hourly, negotiable]
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, top_rated, price_asc, price_desc, most_popular]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated service list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedServices"

    post:
      operationId: createService
      tags: [services]
      summary: Create a new service (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/featured:
    get:
      operationId: getFeaturedServices
      tags: [services]
      summary: Get featured/sponsored services
      parameters:
        - name: wilaya_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Featured services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceCard"
  /services/trending:
    get:
      operationId: getTrendingServices
      tags: [services]
      summary: Get trending services nationally this week
      parameters:
        - name: wilaya_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Trending services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceCard"
  /services/recent:
    get:
      operationId: getRecentServices
      tags: [services]
      summary: Get most recently added services
      parameters:
        - name: wilaya_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Recent services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedServices"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "404":
          description: Not found

    patch:
      operationId: updateService
      tags: [services]
      summary: Update service (seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service (seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SEARCH ───
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, wilaya_id, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [client, seller, admin]
        wilaya_id:
          type: integer
        wilaya_name_fr:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        verified_at:
          type: ["string", "null"]
        banned_at:
          type: ["string", "null"]
        language_preference:
          type: string
          enum: [fr, ar, en]
          default: fr
        trust_score:
          type: ["number", "null"]
        referral_code:
          type: ["string", "null"]
        created_at:
          type: string
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            services_count:
              type: integer
            avg_rating:
              type: ["number", "null"]
            reviews_count:
              type: integer
            completed_orders:
              type: integer
            trust_label:
              type: ["string", "null"]
            badges:
              type: array
              items:
                $ref: "#/components/schemas/Badge"
...
    Badge:
      type: object
      required: [id, name_fr, icon, color]
      properties:
        id:
          type: integer
        name_fr:
          type: string
        name_ar:
          type: ["string", "null"]
        icon:
          type: string
        color:
          type: string

    # ─── CATEGORY ───
...
    ServiceCard:
      type: object
      required: [id, title_fr, price, price_type, delivery_days, seller_id, seller_name, wilaya_id, wilaya_name_fr, category_id, status, created_at]
      properties:
        id:
          type: integer
        title_fr:
          type: string
        title_ar:
          type: ["string", "null"]
        price:
          type: number
        price_type:
          type: string
          enum: [fixed, hourly, negotiable]
        delivery_days:
          type: integer
        seller_id:
          type: integer
        seller_name:
          type: string
        seller_avatar:
          type: ["string", "null"]
        seller_verified:
          type: boolean
        seller_trust_score:
          type: ["number", "null"]
        wilaya_id:
          type: integer
        wilaya_name_fr:
          type: string
        category_id:
          type: integer
        category_name_fr:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
        is_featured:
          type: boolean
        avg_rating:
          type: ["number", "null"]
        reviews_count:
          type: integer
        views_count:
          type: integer
        created_at:
          type: string
    ServiceDetail:
      allOf:
        - $ref: "#/components/schemas/ServiceCard"
        - type: object
          properties:
            description_fr:
              type: ["string", "null"]
            description_ar:
              type: ["string", "null"]
            seller:
              $ref: "#/components/schemas/UserProfile"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    Service:
      allOf:
        - $ref: "#/components/schemas/ServiceCard"
        - type: object
          properties:
            description_fr:
              type: ["string", "null"]
            description_ar:
              type: ["string", "null"]
            rejection_reason:
              type: ["string", "null"]
    ServiceInput:
      type: object
      required: [title_fr, category_id, price, price_type, delivery_days]
      properties:
        title_fr:
          type: string
          minLength: 5
        title_ar:
          type: ["string", "null"]
        description_fr:
          type: ["string", "null"]
        description_ar:
          type: ["string", "null"]
        category_id:
          type: integer
        price:
          type: number
          minimum: 0
        price_type:
          type: string
          enum: [fixed, hourly, negotiable]
        delivery_days:
          type: integer
          minimum: 1
        images:
          type: array
          items:
            type: string
    ServiceUpdate:
      type: object
      properties:
        title_fr:
          type: ["string", "null"]
        title_ar:
          type: ["string", "null"]
        description_fr:
          type: ["string", "null"]
        description_ar:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        price_type:
          type: ["string", "null"]
        delivery_days:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
    PaginatedServices:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCard"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── SEARCH ───
...
    Review:
      type: object
      required: [id, order_id, client_id, seller_id, rating, created_at]
      properties:
        id:
          type: integer
        order_id:
          type: integer
        client_id:
          type: integer
        seller_id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        body:
          type: ["string", "null"]
        is_flagged:
          type: boolean
        seller_reply:
          type: ["string", "null"]
        reply_at:
          type: ["string", "null"]
        client_name:
          type: ["string", "null"]
        client_avatar:
          type: ["string", "null"]
        created_at:
          type: string
...